:arrow_up: django1.9
I had to remove the template_utils library since it was causing
an error. We were using this library to display the feeds, and
I found another library - django-feedparser, that we can use with
django 1.9.

To see the changes, do `git log -p plexus/`... because everything
in the media directory, except for media/css, media/js, and media/img
has changed since I ran collectstatic to get the new django admin
view.