create model for GrainLog
I have Salt running a periodic job that collects all the grain info from
all servers. The plan is to add an endpoint to Plexus that it can submit
that data to. It can then be parsed, matched up to servers and apps, and
Plexus can do a better job of showing the actual current state of
systems.

This adds the very first step, setting up a model that will hold the
dumps. The idea is that the cron job runs periodically and we just store
the last N dumps in Plexus. The 'payload' field is where the json will
get dumped. The 'sha1' field will hold a hash of the payload, which we
can then use to avoid storing duplicates (ie, if the grain info hasn't
changed since the last upload, no need to add a new entry).