add model for leases
Jonah and I were discussing the idea of app leases for quite a while and
it's come up again recently, so I figured it's time to add them.

The idea is that each application we run will have a "lease" that
expires at some point (probably a default of 1 year). A little before
that point, Plexus should send out an email to the maintainer/owner
letting them know that it's going to expire. At that point, the project
team, probably the developer, PM, and a representative from senior
management need to check in with each other and with the client and
positively affirm that the application is still in use and worth keeping
around. At that point, a new lease can be issued and the clock counts
down again. If a lease isn't renewed, we operate on the assumption that
the application is no longer in use and we can begin the process of
sunsetting it.

This PR just adds a super simple base model for the lease. Nothing else
around it is implemented. I'll probably add that functionality in behind
a feature flag so we can refine it and work out exactly how we'd like it
to work. Getting the model in place first just makes that a little simpler.