simple lease UI implemention
This is clearly just some early stage work so we can get started, but it
lets us add a Lease to an Application and view any current and past
leases on the application.

I'll have more to write later on how this is going to work in terms of
project process, but basically, each application that we are running in
production gets a lease period (6 months, a year, 2 years, something
like that). At the end of that period, the dev(s), LD/PM, client, and
senior management have an opportunity to check in, establish that the
application is still being used and that the cost of maintaining it is
worth the value that it provides.

UI work to be done later:

* page to show applications with expired and/or soon to expire leases
* calendar/date-picker widget for end-date
* ability to renew a lease before it expires
* automatic notification on upcoming lease expiration