add field to Server for EC2 instance ID
When AWS sends us notifications about problems with servers, they only
reference them by the Instance ID. Currently, we have to login to the
AWS console to figure out which server they are referring to.

We can simplify this by putting it into Plexus and making it searchable.

This just adds the field. UI for adding/displaying/updating will come
next.