Fix Makefile flow for production builds
Some applications do not have any production npm dependencies. In this case, the build fails as it tries to create the sentinal file in the non-existant node_modules directory. This change always creates a node_modules directory to workaround this issue. I'll migrate the change to the canonical js.mk if this works.